--- src/main/java/gov/nasa/pds/registry/mgr/cfg/RegistryCfg.java
@@ -1,13 +1,0 @@
-package gov.nasa.pds.registry.mgr.cfg;
-
-
-/**
- * Registry (elasticsearch) configuration model.
- * @author karpenko
- */
-public class RegistryCfg
-{
-    public String url;
-    public String indexName;
-    public String authFile;
-}
