+++ src/main/java/gov/nasa/pds/registry/mgr/cmd/dd/ListDDCmd.java
@@ -5,10 +5,10 @@ import java.util.List;
 
 import org.apache.commons.cli.CommandLine;
 
+import gov.nasa.pds.registry.common.cfg.RegistryCfg;
 import gov.nasa.pds.registry.common.es.dao.dd.DataDictionaryDao;
 import gov.nasa.pds.registry.common.es.dao.dd.LddInfo;
 import gov.nasa.pds.registry.mgr.Constants;
-import gov.nasa.pds.registry.mgr.cfg.RegistryCfg;
 import gov.nasa.pds.registry.mgr.cmd.CliCommand;
 import gov.nasa.pds.registry.mgr.dao.RegistryManager;
