+++ src/main/java/gov/nasa/pds/registry/mgr/dao/RegistryManager.java
@@ -4,11 +4,11 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.elasticsearch.client.RestClient;
 
+import gov.nasa.pds.registry.common.cfg.RegistryCfg;
 import gov.nasa.pds.registry.common.es.client.EsClientFactory;
 import gov.nasa.pds.registry.common.es.dao.dd.DataDictionaryDao;
 import gov.nasa.pds.registry.common.util.CloseUtils;
-import gov.nasa.pds.registry.mgr.cfg.RegistryCfg;
-import gov.nasa.pds.registry.mgr.dao.schema.SchemaDao;
+import gov.nasa.pds.registry.common.es.dao.schema.SchemaDao;
 
 
 /**
